// TO_MIGRATE_TO_R-D-LABS/vendors/react/formular.components/src/components/__v2/spinner/spinner.types.ts
import { ComponentSizeType, ComponentVariantType } from 'formular.design.system'

export interface ISpinnerOverriders {
    mainSize?: number
    frameSize?: number
    strokeSize?: number
    strokeOpacity?: number
    strokeColor?: string
    activeColor?: string
}

export interface ISpinnerProps extends ISpinnerOverriders, Partial<React.ComponentProps<'span'>> {
    size?: ComponentSizeType
    color?: ComponentVariantType
    className?: string
    overriders?: ISpinnerOverriders
}